// src/app/interface/tipo.ministerio.interface.ts
// Generated by https://quicktype.io

export interface ResultTipoMinisterios {
  ok:             boolean;
  msg:            string;
  tipoMinisterio: TipoMinisterio[];
}
export interface ResultTipoMinisterio {
  ok:             boolean;
  msg:            string;
  tipoMinisterio: TipoMinisterio;
}
export interface TipoMinisterio {
  id:            number;
  titulo:        string;
  nom_fecha:     string;
  descripcion:   string;
  id_ministerio: number;
  Ministerio:    Ministerio;
}

export interface Ministerio {
  id:     number;
  nombre: string;
  titulo: string;
  logo:   string;
  pdf:    string;
}
export interface TipoMinisterioForm{
  titulo:string,
  nom_fecha:string,
  descripcion:string,
  id_ministerio:string
}